refactor: organize webcomponents into categorized folders
Reorganize 36 component files from a flat directory into 7 subdirectories based on their purpose for better maintainability and navigation. Updated all 50+ import references across component files and external entrypoints.
