+++ packages/core/src/client/standalone/App.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import type { DocksContext } from '@vitejs/devtools-kit/client'
 import { CLIENT_CONTEXT_KEY, getDevToolsRpcClient } from '@vitejs/devtools-kit/client'
-import DockStandalone from '../webcomponents/components/DockStandalone.vue'
+import DockStandalone from '../webcomponents/components/dock/DockStandalone.vue'
 import { createDocksContext } from '../webcomponents/state/context'
 
 const rpc = await getDevToolsRpcClient()
