+++ packages/core/src/client/webcomponents/components/DockEmbedded.ts
@@ -2,7 +2,7 @@ import type { DocksContext } from '@vitejs/devtools-kit/client'
 import type { VueElementConstructor } from 'vue'
 import { defineCustomElement } from 'vue'
 import css from '../.generated/css'
-import Component from './DockEmbedded.vue'
+import Component from './dock/DockEmbedded.vue'
 
 export const DockEmbedded = defineCustomElement(
   Component,
