+++ packages/core/src/client/webcomponents/components/DockStandalone.ts
@@ -2,7 +2,7 @@ import type { DocksContext } from '@vitejs/devtools-kit/client'
 import type { VueElementConstructor } from 'vue'
 import { defineCustomElement } from 'vue'
 import css from '../.generated/css'
-import Component from './DockStandalone.vue'
+import Component from './dock/DockStandalone.vue'
 
 export const DockStandalone = defineCustomElement(
   Component,
