+++ packages/core/src/client/webcomponents/components/dock/DockEntries.vue
