+++ packages/core/src/client/webcomponents/components/dock/DockOverflowButton.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
 import type { DevToolsDockEntry } from '@vitejs/devtools-kit'
 import type { DocksContext } from '@vitejs/devtools-kit/client'
-import type { DevToolsDockEntriesGrouped } from '../state/dock-settings'
+import type { DevToolsDockEntriesGrouped } from '../../state/dock-settings'
 import { watchDebounced } from '@vueuse/core'
 import { computed, h, ref, useTemplateRef } from 'vue'
-import { setDocksOverflowPanel, useDocksOverflowPanel } from '../state/floating-tooltip'
+import { setDocksOverflowPanel, useDocksOverflowPanel } from '../../state/floating-tooltip'
 import DockEntriesWithCategories from './DockEntriesWithCategories.vue'
 import DockEntry from './DockEntry.vue'
