+++ packages/core/src/client/webcomponents/components/dock/DockPanelResizer.vue
