+++ packages/core/src/client/webcomponents/components/dock/DockStandalone.vue
@@ -1,13 +1,13 @@
 <script setup lang="ts">
 import type { DocksContext } from '@vitejs/devtools-kit/client'
 import { computed, markRaw, ref, useTemplateRef, watch } from 'vue'
-import { PersistedDomViewsManager } from '../utils/PersistedDomViewsManager'
+import { PersistedDomViewsManager } from '../../utils/PersistedDomViewsManager'
+import ToastOverlay from '../display/ToastOverlay.vue'
+import FloatingElements from '../floating/FloatingElements.vue'
+import VitePlus from '../icons/VitePlus.vue'
+import ViewBuiltinClientAuthNotice from '../views-builtin/ViewBuiltinClientAuthNotice.vue'
+import ViewEntry from '../views/ViewEntry.vue'
 import DockEntriesWithCategories from './DockEntriesWithCategories.vue'
-import FloatingElements from './FloatingElements.vue'
-import VitePlus from './icons/VitePlus.vue'
-import ToastOverlay from './ToastOverlay.vue'
-import ViewBuiltinClientAuthNotice from './ViewBuiltinClientAuthNotice.vue'
-import ViewEntry from './ViewEntry.vue'
 
 const props = defineProps<{
   context: DocksContext
