+++ packages/core/src/client/webcomponents/components/log/LogItemConstants.ts
