+++ packages/core/src/client/webcomponents/index.ts
@@ -1,3 +1,4 @@
 export * from './components/DockEmbedded'
+export * from './components/DockStandalone'
 export * from './state/docks'
 export * from './utils/PersistedDomViewsManager'
